Add inheritance tests to ili2gpkg
Omitted tests that use an sql setup script that does not work on SQLite
without modifications.
